+++ internal/cli/cli.go
@@ -203,7 +203,6 @@ func NewRootCmd(build BuildInfo) *cobra.Command {
 	root.AddCommand(newTopicsCmd())
 	root.AddCommand(newTagPrepCmd())
 	root.AddCommand(newTagWriteCmd())
-	root.AddCommand(newTagQueueCmd())
 	archiveCmd := newArchiveCmd()
 	archiveCmd.AddCommand(newArchiveInitCmd())
 	archiveCmd.AddCommand(newArchivePushCmd())